# Use Vec::extend in SmallVec::extend when applicable