@Mark-Simulacrum 2018-07-30
Seems good. If you're interested, it'd be great to write some benchmarks for smallvec to have more confidence in such changes.

@bors r+