@ljedrz 2018-07-30
Yes, I was just about to write the same thing; I forgot the lower bound is not to be trusted.

I think adding `arr.reserve(1)` to the loop will fix this. I'll test it shortly.