@Mark-Simulacrum 2018-07-31
Can you push those benchmarks as another commit? That way they will be easier for others to reuse.