@ljedrz 2018-07-31
Sure; are they fine as an extension of the existing `small_vec.rs` file (a test module like in the gist)?