@Mark-Simulacrum 2018-07-31
Yes, that should work.